Made some changes regarding this edgeR change, per Gordon Smyth:
We no longer permit users to pass an offset matrix as an argument when 
running estimateGLMCommonDisp() on an DGEList object.  Hence

   estimateGLMCommonDisp(y, design, offset=offset)

will give an error if y is a DGEList object.  We require the offset matrix 
to already be in y, e.g.,

   y$offset <- offset
   estimateGLMCommonDisp(y, design)

This was always our intended syntax, but now we make it compulsory.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/msmsTests@83865 bc3139a8-67e5-0310-9ffc-ced21a209358